+++ src/test/java/com/iexec/sms/admin/AdminServiceTests.java
@@ -16,33 +16,52 @@
 
 package com.iexec.sms.admin;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mockito;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.boot.test.system.CapturedOutput;
+import org.springframework.boot.test.system.OutputCaptureExtension;
+import org.springframework.context.annotation.Bean;
 
+import javax.sql.DataSource;
 import java.io.File;
+import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 
+@ExtendWith(OutputCaptureExtension.class)
 class AdminServiceTests {
 
-    private final AdminService adminService = new AdminService("", "", "");
+    @Bean
+    public DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .url("jdbc:h2:mem:test")
+                .username("sa")
+                .password("")
+                .build();
+    }
 
+    private final AdminService adminService = new AdminService("jdbc:h2:mem:test", "sa", "", "/tmp/");
     @TempDir
     File tempStorageLocation;
 
     // region backup
     @Test
     void shouldReturnTrueWhenAllParametersAreValid() {
-        AdminService adminServiceSpy = Mockito.spy(adminService);
-
-        Mockito.doReturn(true).when(adminServiceSpy).databaseDump(any());
-        assertTrue(adminServiceSpy.createDatabaseBackupFile(tempStorageLocation.getPath(), "backup.sql"));
+        assertTrue(adminService.createDatabaseBackupFile(tempStorageLocation.getPath(), "backup.sql"));
     }
+
 
     @Test
     void shouldReturnFalseWhenAllParametersAreValidButBackupFailed() {
@@ -96,6 +115,91 @@ class AdminServiceTests {
     @Test
     void shouldReturnNotImplementedWhenCallingRestore() {
         assertEquals("restoreDatabaseFromBackupFile is not implemented", adminService.restoreDatabaseFromBackupFile("", ""));
+    }
+    // endregion
+
+    // region delete-backup
+    @Test
+    void shouldDeleteBackup(CapturedOutput output) throws IOException {
+        final String backupFileName = "backup.sql";
+        final Path tmpFile = Files.createFile(tempStorageLocation.toPath().resolve(backupFileName));
+        assertAll(
+                () -> assertTrue(adminService.deleteBackupFileFromStorage(tempStorageLocation.getPath(), backupFileName)),
+                () -> assertFalse(tmpFile.toFile().exists()),
+                () -> assertTrue(output.getOut().contains("Successfully deleted backup"))
+        );
+    }
+
+    @Test
+    void shouldFailedDeleteWithBackupFileMissing(CapturedOutput output) {
+        final String backupFileName = "backup.sql";
+        assertAll(
+                () -> assertFalse(adminService.deleteBackupFileFromStorage(tempStorageLocation.getPath(), backupFileName)),
+                () -> assertFalse(adminService.deleteBackupFileFromStorage(tempStorageLocation.getPath(), "")),
+                () -> assertFalse(adminService.deleteBackupFileFromStorage("", backupFileName)),
+                () -> assertTrue(output.getOut().contains("Backup file does not exist"))
+        );
+    }
+
+    @Test
+    @Disabled("Can't lock file on linux")
+    void shouldFailedDeleteWithBackupFileCantDelete(CapturedOutput output) throws IOException {
+        final String backupFileName = "backupLock.sql";
+        final Path tmpFile = Files.createFile(tempStorageLocation.toPath().resolve(backupFileName));
+
+
+        try (FileChannel ignored = FileChannel.open(tmpFile, StandardOpenOption.WRITE)) {
+            assertAll(
+                    () -> assertFalse(adminService.deleteBackupFileFromStorage(tempStorageLocation.getPath(), backupFileName))
+            );
+        }
+    }
+
+    @Test
+    void shouldFailedToDeleteWithBackupFileOutOfStorage(CapturedOutput output) {
+        assertAll(
+                () -> assertFalse(adminService.deleteBackupFileFromStorage("/backup", "backup.sql")),
+                () -> assertTrue(output.getOut().contains("Backup file is outside of storage file system"))
+        );
+    }
+
+    // endregion
+
+    //region utils
+
+    @Test
+    void testIsPathInBaseDirectory() {
+        assertAll(
+                () -> assertFalse(adminService.isPathInBaseDirectory("/tmp/../../backup.sql")),
+                () -> assertTrue(adminService.isPathInBaseDirectory("/tmp/backup.sql")),
+                () -> assertTrue(adminService.isPathInBaseDirectory("/tmp/backup-copy.sql"))
+        );
+    }
+
+    @Test
+    void testNormalizePathWithSeparator() {
+        String path1 = "test/path/";
+        String path2 = "test/path";
+        assertAll(
+                () -> assertEquals(path1, adminService.normalizePathWithSeparator(path1)),
+                () -> assertEquals(path1, adminService.normalizePathWithSeparator(path2))
+        );
+    }
+
+    @Test
+    void testCommonsParametersValidation() {
+        // Valid case
+        final String validStorageLocation = "test/path/";
+        final String validBackupFileName = "backup.sql";
+        final String emptyStorageLocation = "";
+        final String emptyBackupFileName = "";
+
+        assertAll(
+                () -> assertTrue(adminService.commonsParametersValidation(validStorageLocation, validBackupFileName)),
+                () -> assertFalse(adminService.commonsParametersValidation(emptyStorageLocation, validBackupFileName)),
+                () -> assertFalse(adminService.commonsParametersValidation(validStorageLocation, emptyBackupFileName)),
+                () -> assertFalse(adminService.commonsParametersValidation(emptyStorageLocation, emptyBackupFileName))
+        );
     }
     // endregion
 }