+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Create admin endpoints foundation. (#208 #209)
 - Add H2 database connection informations and storage ID decoding method. (#210)
 - Add the ability to trigger a backup via a dedicated endpoint. (#211)
+- Add the ability to trigger a delete via a dedicated endpoint. (#)
 
 ## [[8.3.0]](https://github.com/iExecBlockchainComputing/iexec-sms/releases/tag/v8.3.0) 2023-09-28